frontend/requirements.txt
flask==2.0.1
flask-login==0.5.0
gunicorn==20.1.0
wtforms==2.3.3
pyjwt==2.1.0
requests==2.25.1
google-auth==1.31.0
beautifulsoup4==4.9.3